requirements.txt
openai
chromadb
pyyaml
python-dotenv
streamlit
pyttsx3
pydub>=0.25.1
Pillow>=10.0.0
faster-whisper>=1.0.0
